+++ resvg-qt/qt-bindings/qt_capi.cpp
@@ -42,6 +42,28 @@ qtc_qimage * qtc_qimage_create_rgba(uint32_t width, uint32_t height)
     return reinterpret_cast<qtc_qimage*>(img);
 }
 
+qtc_qimage* qtc_qimage_from_file(const char *path)
+{
+    QImage *img = new QImage(QString::fromUtf8(path));
+
+    if (img->isNull()) {
+        return 0;
+    }
+
+    return reinterpret_cast<qtc_qimage*>(img);
+}
+
+qtc_qimage* qtc_qimage_from_data(const uint8_t *data, int size)
+{
+    QImage img = QImage::fromData(data, size);
+
+    if (img.isNull()) {
+        return 0;
+    }
+
+    return reinterpret_cast<qtc_qimage*>(new QImage(img));
+}
+
 uint8_t* qtc_qimage_get_data(qtc_qimage *c_img)
 {
     return IMAGE_CAST->bits();