+++ resvg-qt/src/filter.rs
@@ -807,9 +807,7 @@ fn apply_image(
             if format == usvg::ImageFormat::SVG {
                 super::image::draw_svg(data, view_box, opt, &mut p);
             } else {
-                super::image::draw_raster(
-                    format, data, view_box, fe.rendering_mode, opt, &mut p
-                );
+                super::image::draw_raster(data, view_box, fe.rendering_mode, opt, &mut p);
             }
         }
         usvg::FeImageKind::Use(ref id) => {